Implement search history
Router now also handles /search/:query. This query is passed in as a
prop to <Search>. If the query prop is present, a `search()` function is
called that fires an ajax GET to API.

Since the ajax requesting functionality is refactored out to `search()`,
`handleSubmit` now simply redirects to a new url corresponding with the
entered query. At this point, `componentWillMount` hook will be
activated, calling the `search()` function again.