make scrollbar always visible
When a <Chapter> is fetching pages, its vertical length typically does
not exceed the viewport's height (because only a "loading" icon is
displayed). When it's finished fetching and displaying pages, it will
exceed the viewport's height, which causes a vertical scrollbar to
appear. This changes the viewport's width, which causes the page's
components to move a little to adjust to that change. This might make
the whole page seem to be flickering. Making the scrollbar always appear makes
sure there's no movement whatsoever.