fix fucking stupid is_fresh() logic bug
FUCK FUCK FUCK FUCK FUCK WHY THE BLEEDING FUCK WAS I SO FUCKING
STUPID?????????

On a serious note: a timedelta's `seconds` attribute is not the total
number of seconds of the whole duration, but the remaining seconds after
deducting out complete days. This fixes the bug where sometimes (more
often than not) a series' info is not updated because it is incorrectly
considered "fresh".

Again, what the fuck was I thinking?